+++ src/Profile/Magento/Converter/OrderConverter.php
@@ -83,7 +83,7 @@ abstract class OrderConverter extends MagentoConverter
         MagentoMappingServiceInterface $mappingService,
         LoggingServiceInterface $loggingService,
         TaxCalculator $taxCalculator,
-        NumberRangeValueGeneratorInterface $numberRangeValueGenerator
+        NumberRangeValueGeneratorInterface $numberRangeValueGenerator,
     ) {
         parent::__construct($mappingService, $loggingService);
 
@@ -625,9 +625,6 @@ abstract class OrderConverter extends MagentoConverter
 
         if (isset($originalData['region_id'])
             && isset($originalData['region_code'])
-            && isset($originalData['country_iso2'])
-            && isset($originalData['country_iso3'])
-            && $this->mappingService instanceof MagentoMappingServiceInterface
         ) {
             $countryStateUuid = $this->mappingService->getCountryStateUuid(
                 $originalData['region_id'],