+++ src/Profile/Magento/Premapping/PaymentMethodReader.php
@@ -46,7 +46,7 @@ abstract class PaymentMethodReader extends AbstractPremappingReader
      */
     public function __construct(
         GatewayRegistryInterface $gatewayRegistry,
-        EntityRepository $paymentMethodRepo
+        EntityRepository $paymentMethodRepo,
     ) {
         $this->gatewayRegistry = $gatewayRegistry;
         $this->paymentMethodRepo = $paymentMethodRepo;