+++ src/Profile/Magento2/Gateway/Local/Magento2LocalGateway.php
@@ -50,7 +50,7 @@ abstract class Magento2LocalGateway implements MagentoGatewayInterface
         EnvironmentReaderInterface $localEnvironmentReader,
         TableReaderInterface $localTableReader,
         ConnectionFactoryInterface $connectionFactory,
-        EntityRepository $currencyRepository
+        EntityRepository $currencyRepository,
     ) {
         $this->readerRegistry = $readerRegistry;
         $this->localEnvironmentReader = $localEnvironmentReader;